[type] 타입 분리
