+++ src/context/notification/index.ts
@@ -1,11 +1,5 @@
 import { createContext, useContext } from "react";
-
-export interface Notification {
-  id: number;
-  message: string;
-  type: "info" | "success" | "warning" | "error";
-}
-
+import { Notification } from "../../types/notification";
 export interface NotificationContextType {
   notifications: Notification[];
   addNotification: (message: string, type: Notification["type"]) => void;
